Only setup if keys and LAZYREMOVE

+++ code/game/objects/items/keyrings.dm
@@ -19,12 +19,15 @@
 
 /obj/item/storage/keyring/Initialize()
 	. = ..()
+	if(!length(src.keys))
+		return
 	if(length(src.keys) > 10)
 		warning("Keyring [src] has too many keys and the list will get cut short!")
 	for(var/X as anything in keys)
 		var/obj/item/key/new_key = new X(loc)
 		if(!SEND_SIGNAL(src, COMSIG_TRY_STORAGE_INSERT, new_key, null, TRUE, FALSE))
 			qdel(new_key)
+		LAZYREMOVE(src.keys, X)
 
 	update_icon()
 	update_desc()
